Refactored graph importing methods to ViewModel
