+++ app/src/main/java/dk/dtu/group22/beeware/view/GraphViewModel.java
@@ -2,8 +2,18 @@ package dk.dtu.group22.beeware.view;
 
 import androidx.lifecycle.ViewModel;
 
+import com.github.mikephil.charting.data.Entry;
+
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
+
+import dk.dtu.group22.beeware.data.entities.Hive;
+import dk.dtu.group22.beeware.data.entities.Measurement;
+
 public class GraphViewModel extends ViewModel {
-    // Serves as a temporary "state" for when the GraphActivity is redrawn.
+
+    // State
 
     private boolean weightLineVisible = true, temperatureLineVisible = false,
             sunlightLineVisible = false, humidityLineVisible = false;
@@ -22,11 +32,11 @@ public class GraphViewModel extends ViewModel {
         this.xCenter = xCenter;
     }
 
-    public float getDefaultZoom() {
+    public float getZoom() {
         return defaultZoom;
     }
 
-    public void setDefaultZoom(float defaultZoom) {
+    public void setZoom(float defaultZoom) {
         this.defaultZoom = defaultZoom;
     }
 
@@ -60,5 +70,69 @@ public class GraphViewModel extends ViewModel {
 
     public void setHumidityLineVisible(boolean humidityLineVisible) {
         this.humidityLineVisible = humidityLineVisible;
+    }
+
+    // Data handling for graph
+
+
+    public List<Entry> extractWeight(Hive hive) {
+        List<Entry> res = new ArrayList<>();
+        for (Measurement measure : hive.getMeasurements()) {
+            float time = (float) measure.getTimestamp().getTime();
+            float weight = (float) measure.getWeight();
+            res.add(new Entry(time, weight));
+        }
+        return res;
+    }
+
+    public List<Entry> extractTemperature(Hive hive) {
+        List<Entry> res = new ArrayList<>();
+        for (Measurement measure : hive.getMeasurements()) {
+            float time = (float) measure.getTimestamp().getTime();
+            float temp = (float) measure.getTempIn();
+            res.add(new Entry(time, temp));
+        }
+        return res;
+    }
+
+    public List<Entry> extractIlluminance(Hive hive) {
+        List<Entry> res = new ArrayList<>();
+        for (Measurement measure : hive.getMeasurements()) {
+            float time = (float) measure.getTimestamp().getTime();
+            float illum = (float) measure.getIlluminance();
+            res.add(new Entry(time, illum));
+        }
+        return res;
+    }
+
+    public List<Entry> extractHumidity(Hive hive) {
+        List<Entry> res = new ArrayList<>();
+        for (Measurement measure : hive.getMeasurements()) {
+            float time = (float) measure.getTimestamp().getTime();
+            float humid = (float) measure.getHumidity();
+            res.add(new Entry(time, humid));
+        }
+        return res;
+    }
+
+    /**
+     * @param hive
+     * @param start
+     * @param end
+     * @return A list containing measurements, in correct order, from start to end.
+     */
+    // Filters time intervals from a list,
+    // Only picks midnight when time interval is greater than 24hours
+    public List<Measurement> filterTimeInterval(Hive hive, Timestamp start, Timestamp end) {
+        //TODO: Optimize with binary search.
+        List<Measurement> res = new ArrayList<>();
+        for (Measurement measure : hive.getMeasurements()) {
+            Timestamp t = measure.getTimestamp();
+            // start <= t <= end
+            if (start.compareTo(t) <= 0 && t.compareTo(end) <= 0) {
+                res.add(measure);
+            }
+        }
+        return res;
     }
 }
